% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{clean_data}
\alias{clean_data}
\title{This cleans the data and exports them into a new folder}
\usage{
clean_data(fs)
}
\arguments{
\item{fs}{A flowset}
}
\value{
Cleaned data
}
\description{
This cleans the data and exports them into a new folder
}
\examples{
clean_data(fs)
# When you need to re-initialize your clean data:
resQC <- flowCore::read.flowSet(path = "resQC/", pattern = ".*fcs")
}
