% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{compensate_data}
\alias{compensate_data}
\title{Compensate the Data}
\usage{
compensate_data(comp_controls, channels, us_num)
}
\arguments{
\item{comp_controls}{the folder containing single-stained
compensation controls}

\item{channels}{the channels to be compensated in the
pattern 'BV421-A|BV570-A'}

\item{us_num}{the number of the unstained control tube in
the set}
}
\value{
a list with a summary of the comp controls and a
spillover matrix
}
\description{
It will attempt to find the compensation
controls using the regex '.*Setup.*fcs'
}
\examples{
gs.uncomped <- GatingSet(resQC)
spillover_matrix <- compensate_data(comp_controls, 'BV421-A|BV570-A', 3)[[2]]
gs.comped <- compensate(gs.uncomped, spillover_matrix)
}
