% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\docType{data}
\name{cs}
\alias{cs}
\title{A clean cytoset for practice}
\format{
A cytoset
}
\source{
Rekha Garg, Natalie Silver
}
\usage{
cs
}
\description{
Just something I am borrowing from an experiment.
}
\keyword{datasets}
