% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fs}
\alias{fs}
\title{Example data.frame}
\format{
A flowset
\describe{
\item{x}{some numbers}
\item{y}{some other numbers}
}
}
\usage{
fs
}
\description{
Example data.frame
}
\keyword{datasets}
