% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{generate_flowset}
\alias{generate_flowset}
\title{This QCs all of the files in your folder}
\usage{
generate_flowset(folder)
}
\arguments{
\item{folder}{The location of your FCS files}
}
\value{
A flowset
}
\description{
This QCs all of the files in your folder
}
