% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{get_fcs_resultsQC}
\alias{get_fcs_resultsQC}
\title{Load your QC results or, if none, run QC on your data}
\usage{
get_fcs_resultsQC(QC_folder = "./resultsQC/", raw_folder = "./rawData/")
}
\arguments{
\item{QC_folder}{the path to the folder with the QCed data}

\item{raw_folder}{the path to the folder with the raw files}
}
\value{
a cytoset with QCed data
}
\description{
Load your QC results or, if none, run QC on your data
}
\examples{
\dontrun{
get_fcs_resultsQC(QC_folder = "./RG Microbiome -APCs panel/resultsQC/",
                  raw_folder = "../20200810NJ-ACEFlow/good/")
}
}
