% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{make_gates_from_fmos}
\alias{make_gates_from_fmos}
\title{Set gates based on fluorescence-minus-one (FMO) controls}
\usage{
make_gates_from_fmos(
  gating_set = gs,
  parent = "SSC-singlet",
  trim = 0,
  search_term = "FMO"
)
}
\arguments{
\item{gating_set}{a gating set}

\item{parent}{the name of the parent population}

\item{trim}{how many to remove from the end of the markernames list (likely necessary because of QC)}

\item{search_term}{what string must this function use to identify the FMO?}
}
\value{
a gating set
}
\description{
Set gates based on fluorescence-minus-one (FMO) controls
}
\examples{
\dontrun{
make_gates_from_fmos <- function(gating_set = gs, parent = "SSC-singlet", trim = 1)
}
}
