% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{quick_gate}
\alias{quick_gate}
\title{Quickly gate your populations}
\usage{
quick_gate(gs)
}
\arguments{
\item{gs}{a gatingset}
}
\value{
the paths generated and a gatingset with gates
}
\description{
Quickly gate your populations
}
