% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{set_scatter_gate}
\alias{set_scatter_gate}
\title{Set a polygon gate}
\usage{
set_scatter_gate(
  gating_set = gs,
  gate_coordinates = c(1000, 150000, 1000, 1e+05, 12000, 1e+05, 12000, 150000),
  gate_name = "scatter",
  dimensions = c("FSC-A", "SSC-A"),
  parent = "root"
)
}
\arguments{
\item{gating_set}{a gating set}

\item{gate_coordinates}{a vector of points in the format "c(x1, y1, x2, y2 . . . " and so on}

\item{gate_name}{the name by which you will call the gate}

\item{dimensions}{the X and Y parameters of the gate (in that order) as strings in a list}

\item{parent}{a string; the name of the parent gate}
}
\value{
modified gating set
}
\description{
Set a polygon gate
}
\examples{
\dontrun{
set_scatter_gate(gating_set = gs,
                 gate_coordinates = c(2.0e+04, 2.0e+04,
                                      1.4e+04, 1.5e+04,
                                      1.4e+04, 0.2e+04,
                                      3.0e+04, 0.2e+04,
                                      3.0e+04, 2.0e+04),
                 gate_name = "scatter")
}
}
