% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{test_plot}
\alias{test_plot}
\title{Try your gate before setting your gate}
\usage{
test_plot(
  gating_set = gs,
  gate_coordinates = c(1000, 150000, 1000, 1e+05, 12000, 1e+05, 12000, 150000),
  parent = "root",
  dimensions = list("FSC-A", "SSC-A")
)
}
\arguments{
\item{gating_set}{a gating set}

\item{gate_coordinates}{a vector of points in the format "c(x1, y1, x2, y2 . . . " and so on}

\item{parent}{a string; the name of the parent gate}

\item{dimensions}{the X and Y parameters of the gate (in that order) as strings in a list}
}
\value{
a plot
}
\description{
Try your gate before setting your gate
}
\examples{
\dontrun{
set_scatter_gate(gating_set = gs,
gate_coordinates = c(2.0e+04, 2.0e+04,
                     1.4e+04, 1.5e+04,
                     1.4e+04, 0.2e+04,
                     3.0e+04, 0.2e+04,
                     3.0e+04, 2.0e+04),
gate_name = "scatter")
}
}
