% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{update_gate}
\alias{update_gate}
\title{Update Gate}
\usage{
update_gate(gs, gate, filter, parent)
}
\arguments{
\item{gs}{a gating set}

\item{gate}{the ID of the gate}

\item{filter}{the new filter object}

\item{parent}{the parent node}
}
\description{
Update Gate
}
