% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{get_fcs_resultsQC}
\alias{get_fcs_resultsQC}
\title{Load your QC results or, if none, run QC on your data}
\usage{
get_fcs_resultsQC(QC_folder = "./resultsQC/", raw_folder = "./rawData/")
}
\arguments{
\item{QC_folder}{the path to the folder with the QCed data}

\item{raw_folder}{the path to the folder with the raw files}
}
\value{
a cytoset with QCed data
}
\description{
Load your QC results or, if none, run QC on your data
}
\details{
Only use this feature if you are in the parent folder of the files
with which you are working. If the filepath is too long or perhapse
has some odd characters, then you read.flowSet() may have some
difficulties.
}
\examples{
\dontrun{
get_fcs_resultsQC(QC_folder = "./RG Microbiome -APCs panel/resultsQC/",
                  raw_folder = "../20200810NJ-ACEFlow/good/")
}
}
